requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
conda==4.3.16
dateparser==1.1.1
distlib==0.3.4
filelock==3.6.0
frozenlist==1.3.0
idna==3.3
lxml==4.9.1
multidict==6.0.2
numpy==1.23.1
pandas==1.4.3
pipenv==2022.4.8
platformdirs==2.5.1
pycosat==0.6.3
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
tradingview-ta==3.2.10
tzdata==2022.1
tzlocal==4.2
ujson==5.4.0
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
websockets==10.3
yarl==1.7.2
